requirements.txt
annotated-types==0.7.0
anyio==4.12.1
boto3==1.42.60
botocore==1.42.60
cffi==2.0.0
click==8.3.1
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
dotenv==0.9.9
fastapi==0.116.1
fonttools==4.61.1
h11==0.16.0
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpld3==0.5.12
narwhals==2.17.0
numpy==2.4.2
packaging==26.0
pandas==3.0.1
pillow==12.1.1
plotly==6.6.0
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.47.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.35.0
